requirements.txt
python-dotenv
requests
langchain
langchain-community
langchain-groq
sentence-transformers
faiss-cpu
numpy
pydantic